requirements.txt
# ========== CORE DEPENDENCIES ==========
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ========== HTTP & APIs ==========
requests==2.31.0
aiohttp==3.9.1
urllib3==2.1.0

# ========== VIDEO PROCESSING ==========
moviepy==1.0.3
imageio==2.33.1
imageio-ffmpeg==1.4.1
numpy==1.24.3
Pillow==10.1.0

# ========== AUDIO ==========
elevenlabs==0.2.15
pydub==0.25.1

# ========== LLMs & AI ==========
google-generativeai==0.3.0
anthropic==0.7.6
pytrends==4.9.2

# ========== GOOGLE SERVICES ==========
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0

# ========== DATA PROCESSING ==========
pandas==2.1.4
scipy==1.11.4

# ========== DATABASE ==========
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ========== LOGGING & CONFIG ==========
python-json-logger==2.0.7

# ========== ASYNC UTILITIES ==========
asyncio-contextmanager==1.0.0

# ========== DEVELOPMENT (Optional) ==========
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.0
flake8==6.1.0
